+++ back/database/migrations/2024_02_22_131702_create_transaction_template_audio_names_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
         Schema::create('transaction_template_audio_names', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->integer("transaction_template_id");
+            $table->unsignedBigInteger("transaction_template_id");
             $table->string("name", 1000);
             $table->foreign('transaction_template_id')->references('id')->on('transaction_templates')->onDelete('cascade');
         });
